refactor(BadgesFromPkg): remove possibility to add badges with options
BREAKING CHANGE: BadgesFromPkg no longer takes a overrideBadges prop.
It now takes a disabledBadges prop which is an array of names of default badges to disable.
It is no longer possible to add badges to BadgesFromPkg that require options.
Instead, manually add the badges to your README file next to BadgesFromPkg.